`fn ac_dump`: Make `w`, `h` args `usize`s.

+++ src/recon_tmpl_16.rs
@@ -3361,7 +3361,7 @@ pub unsafe extern "C" fn dav1d_recon_b_intra_16bpc(
                         pl += 1;
                     }
                     if DEBUG_BLOCK_INFO(&*f, &*t) && 0 != 0 {
-                        ac_dump(ac, 4 * cbw4, 4 * cbh4, "ac");
+                        ac_dump(ac, 4 * cbw4 as usize, 4 * cbh4 as usize, "ac");
                         hex_dump(
                             uv_dst[0],
                             stride,
